\name{rfseries-methods}
\docType{methods}
\alias{rfseries,numeric,numeric,numeric-method}
\title{ Methods for Function \code{rfseries} }
\description{
Generate random formal series rfseries 
}
\section{Methods}{
\describe{

\item{\code{signature(var = "numeric", cf = "numeric", k = "numeric", m = "numeric")}}{
Method is describe in documentation of rfseries function
}
}}
\seealso{
\code{\link{rfseries}}
}

\keyword{methods}

